+++ modelcraft/tests/ccp4/test_coot.py
@@ -1,4 +1,4 @@
-from modelcraft.jobs.coot import Prune, FixSideChains
+from modelcraft.jobs.coot import Prune
 from modelcraft.structure import ModelStats
 from . import insulin_refmac
 
@@ -14,16 +14,3 @@ def test_insulin_prune():
     stats_in = ModelStats(refmac.structure)
     stats_out = ModelStats(coot.structure)
     assert stats_out.residues < stats_in.residues
-
-
-def test_insulin_fix_side_chains():
-    refmac = insulin_refmac()
-    refmac.structure.remove_alternative_conformations()
-    coot = FixSideChains(
-        structure=refmac.structure,
-        fphi_best=refmac.fphi_best,
-        fphi_diff=refmac.fphi_diff,
-    ).run()
-    stats_in = ModelStats(refmac.structure)
-    stats_out = ModelStats(coot.structure)
-    assert stats_out.residues == stats_in.residues