+++ CHANGELOG.md
@@ -3,11 +3,15 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](https://semver.org).
 
-## [4.1.0] - Unreleased
+## [5.0.0] - Unreleased
 
 ### Added
 
 - New validation with a new script: modelcraft-validation
+
+### Removed
+
+- Final Coot side-chain fixing step
 
 ## [4.0.1] - 2024-05-14